feat(timer): Implements the `ehal::delay::DelayNs` trait for `timer::TimerCounter` and refactors it a bit to pull out common code and reduce code duplication between chip variants.
