+++ hal/src/timer_params.rs
@@ -38,10 +38,6 @@ impl TimerParams {
 
         let cycles: u32 = ticks / divider;
 
-        if cycles > u16::MAX as u32 {
-            panic!("cycles {} is out of range for a 16 bit counter", cycles);
-        }
-
         TimerParams {
             divider: divider as u16,
             cycles,
